requirements.txt
torch
einops
numpy
tqdm
sentencepiece
datasets
discord
python-dotenv